stopIndexing

+++ src/main/java/searchengine/repository/LemmaRepository.java
@@ -3,9 +3,10 @@ package searchengine.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import searchengine.model.Lemma;
-import searchengine.model.SiteE;
+
 
 @Repository
-public interface LemmaRepository extends JpaRepository<Lemma, Long> {
+public interface LemmaRepository extends JpaRepository<Lemma, Integer> {
     int countBySiteId(int siteId);
+
 }
